Renaming ContinuousBN into NamedDAG
