+++ lib/src/CMakeLists.txt
@@ -4,22 +4,22 @@ ot_add_current_dir_to_include_dirs ()
 
 ot_add_source_file ( Utils.cxx )
 ot_add_source_file ( MixedHistogramUserDefined.cxx )
+ot_add_source_file ( NamedDAG.cxx)
 ot_add_source_file ( NamedJunctionTree.cxx)
 ot_add_source_file ( StratifiedCache.cxx )
 ot_add_source_file ( ContinuousTTest.cxx )
 ot_add_source_file ( IndicesManip.cxx )
 ot_add_source_file ( ContinuousPC.cxx )
-ot_add_source_file ( ContinuousBayesianNetwork.cxx )
 ot_add_source_file ( JunctionTreeBernsteinCopula.cxx )
 
 ot_install_header_file ( MixedHistogramUserDefined.hxx )
 ot_install_header_file ( Utils.hxx )
+ot_install_header_file ( NamedDAG.hxx )
 ot_install_header_file ( NamedJunctionTree.hxx )
 ot_install_header_file ( StratifiedCache.hxx )
 ot_install_header_file ( ContinuousTTest.hxx )
 ot_install_header_file ( IndicesManip.hxx )
 ot_install_header_file ( ContinuousPC.hxx )
-ot_install_header_file ( ContinuousBayesianNetwork.hxx )
 ot_install_header_file ( JunctionTreeBernsteinCopula.hxx )
 
 include_directories (${INTERNAL_INCLUDE_DIRS})
