+++ lib/test/CMakeLists.txt
@@ -83,12 +83,12 @@ include_directories ( ${INTERNAL_INCLUDE_DIRS} )
 ot_check_test ( Utils_std )
 ot_check_test ( MixedHistogramUserDefined_std )
 ot_check_test ( NamedJunctionTree_std )
+ot_check_test ( NamedDAG_std )
 ot_check_test ( JunctionTreeBernsteinCopula_std )
 ot_check_test ( IndicesManip_std )
 ot_check_test ( StratifiedCache_std )
 ot_check_test ( ContinuousTTest_std )
 ot_check_test ( ContinuousPC_std )
-ot_check_test ( ContinuousBayesianNetwork_std )
 
 add_custom_target ( cppcheck COMMAND ${CMAKE_CTEST_COMMAND} -R "^cppcheck_"
                     DEPENDS ${CHECK_TO_BE_RUN}
